added apps and filters

+++ poultryapi/culling/models.py
@@ -5,8 +5,8 @@ class Culling(models.Model):
     date = models.DateField(blank=True, null=True)
     quantity_sold = models.CharField(max_length=200)    
     weight_per_bird_kg = models.IntegerField()
-    price_per_bird = models.IntegerField()  
-    total_amount = models.IntegerField()
+    price_per_bird = models.FloatField()           
+    total_amount = models.FloatField()
     buyer_name = models.CharField(max_length=100)
     payment_method = models.CharField(max_length=100)
