+++ poultryapi/expenses/models.py
@@ -3,10 +3,10 @@ from django.db import models
 class Expenses(models.Model):      
     expense_type = models.CharField(max_length=200)   
     flock = models.CharField(max_length=200)  
-    amount = models.IntegerField()
+    amount = models.FloatField()
     date = models.DateField(blank=True, null=True)
     description = models.CharField(max_length=200)    
     paid_by = models.CharField(max_length=100)
     payment_method = models.CharField(max_length=100)  
-    
-                                      
+        
+
